Remove useless storage classes
