+++ dbms/src/DataStreams/PushingToViewsBlockOutputStream.cpp
@@ -12,9 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#include <Common/Exception.h>
 #include <DataStreams/PushingToViewsBlockOutputStream.h>
 #include <DataStreams/SquashingBlockInputStream.h>
 #include <Interpreters/InterpreterSelectQuery.h>
+#include <Storages/IStorage.h>
 
 
 namespace DB
@@ -39,26 +41,7 @@ PushingToViewsBlockOutputStream::PushingToViewsBlockOutputStream(
     {
         Dependencies dependencies = context.getDependencies(database, table);
 
-        /// We need special context for materialized views insertions
-        if (!dependencies.empty())
-        {
-            views_context = std::make_unique<Context>(context);
-        }
-
-        for (const auto & database_table : dependencies)
-        {
-            auto dependent_table = context.getTable(database_table.first, database_table.second);
-            const auto & materialized_view = dynamic_cast<const StorageMaterializedView &>(*dependent_table);
-
-            auto query = materialized_view.getInnerQuery();
-            BlockOutputStreamPtr out = std::make_shared<PushingToViewsBlockOutputStream>(
-                database_table.first,
-                database_table.second,
-                dependent_table,
-                *views_context,
-                ASTPtr());
-            views.emplace_back(ViewInfo{std::move(query), database_table.first, database_table.second, std::move(out)});
-        }
+        RUNTIME_CHECK_MSG(dependencies.empty(), "Do not support ClickHouse's materialized view");
     }
 
     /* Do not push to destination table if the flag is set */
@@ -73,37 +56,6 @@ void PushingToViewsBlockOutputStream::write(const Block & block)
 {
     if (output)
         output->write(block);
-
-    /// Insert data into materialized views only after successful insert into main table
-    for (auto & view : views)
-    {
-        try
-        {
-            BlockInputStreamPtr from = std::make_shared<OneBlockInputStream>(block);
-            InterpreterSelectQuery select(view.query, *views_context, {}, QueryProcessingStage::Complete, 0, from);
-            BlockInputStreamPtr in = std::make_shared<MaterializingBlockInputStream>(select.execute().in);
-            /// Squashing is needed here because the materialized view query can generate a lot of blocks
-            /// even when only one block is inserted into the parent table (e.g. if the query is a GROUP BY
-            /// and two-level aggregation is triggered).
-            in = std::make_shared<SquashingBlockInputStream>(
-                in,
-                context.getSettingsRef().min_insert_block_size_rows,
-                context.getSettingsRef().min_insert_block_size_bytes,
-                /*req_id=*/"");
-
-            in->readPrefix();
-
-            while (Block result_block = in->read())
-                view.out->write(result_block);
-
-            in->readSuffix();
-        }
-        catch (Exception & ex)
-        {
-            ex.addMessage("while pushing to view " + view.database + "." + view.table);
-            throw;
-        }
-    }
 }
 
 } // namespace DB
