+++ dbms/src/Storages/Transaction/PartitionStreams.cpp
@@ -38,7 +38,6 @@ namespace DB
 namespace FailPoints
 {
 extern const char pause_before_apply_raft_cmd[];
-extern const char pause_before_apply_raft_snapshot[];
 extern const char force_set_safepoint_when_decode_block[];
 extern const char unblock_query_init_after_write[];
 extern const char pause_query_init[];
@@ -291,7 +290,7 @@ std::optional<RegionDataReadInfoList> ReadRegionCommitCache(const RegionPtr & re
     return data_list_read;
 }
 
-void RemoveRegionCommitCache(const RegionPtr & region, const RegionDataReadInfoList & data_list_read, bool lock_region = true)
+void RemoveRegionCommitCache(const RegionPtr & region, const RegionDataReadInfoList & data_list_read, bool lock_region)
 {
     /// Remove data in region.
     auto remover = region->createCommittedRemover(lock_region);
@@ -394,105 +393,6 @@ RegionTable::ResolveLocksAndWriteRegionRes RegionTable::resolveLocksAndWriteRegi
                           [](auto & r) -> ResolveLocksAndWriteRegionRes { return std::move(r); },
                       },
                       region_data_lock);
-}
-
-/// Pre-decode region data into block cache and remove committed data from `region`
-RegionPtrWithBlock::CachePtr GenRegionPreDecodeBlockData(const RegionPtr & region, Context & context)
-{
-    const auto & tmt = context.getTMTContext();
-    {
-        Timestamp gc_safe_point = 0;
-        if (auto pd_client = tmt.getPDClient(); !pd_client->isMock())
-        {
-            gc_safe_point
-                = PDClientHelper::getGCSafePointWithRetry(pd_client, false, context.getSettingsRef().safe_point_update_interval_seconds);
-        }
-        /**
-         * In 5.0.1, feature `compaction filter` is enabled by default. Under such feature tikv will do gc in write & default cf individually.
-         * If some rows were updated and add tiflash replica, tiflash store may receive region snapshot with unmatched data in write & default cf sst files.
-         */
-        region->tryCompactionFilter(gc_safe_point);
-    }
-    std::optional<RegionDataReadInfoList> data_list_read = std::nullopt;
-    try
-    {
-        data_list_read = ReadRegionCommitCache(region, true);
-        if (!data_list_read)
-            return nullptr;
-    }
-    catch (const Exception & e)
-    {
-        if (e.code() == ErrorCodes::ILLFORMAT_RAFT_ROW)
-        {
-            // br or lighting may write illegal data into tikv, skip pre-decode and ingest sst later.
-            LOG_WARNING(Logger::get(__PRETTY_FUNCTION__),
-                        "Got error while reading region committed cache: {}. Skip pre-decode and keep original cache.",
-                        e.displayText());
-            // set data_list_read and let apply snapshot process use empty block
-            data_list_read = RegionDataReadInfoList();
-        }
-        else
-            throw;
-    }
-
-    TableID table_id = region->getMappedTableID();
-    Int64 schema_version = DEFAULT_UNSPECIFIED_SCHEMA_VERSION;
-    Block res_block;
-
-    const auto atomic_decode = [&](bool force_decode) -> bool {
-        Stopwatch watch;
-        auto storage = tmt.getStorages().get(table_id);
-        if (storage == nullptr || storage->isTombstone())
-        {
-            if (!force_decode) // Need to update.
-                return false;
-            if (storage == nullptr) // Table must have just been GC-ed.
-                return true;
-        }
-
-        /// Get a structure read lock throughout decode, during which schema must not change.
-        TableStructureLockHolder lock;
-        try
-        {
-            lock = storage->lockStructureForShare(getThreadNameAndID());
-        }
-        catch (DB::Exception & e)
-        {
-            // If the storage is physical dropped (but not removed from `ManagedStorages`) when we want to decode snapshot, consider the decode done.
-            if (e.code() == ErrorCodes::TABLE_IS_DROPPED)
-                return true;
-            else
-                throw;
-        }
-
-        DecodingStorageSchemaSnapshotConstPtr decoding_schema_snapshot;
-        std::tie(decoding_schema_snapshot, std::ignore) = storage->getSchemaSnapshotAndBlockForDecoding(lock, false);
-        res_block = createBlockSortByColumnID(decoding_schema_snapshot);
-        auto reader = RegionBlockReader(decoding_schema_snapshot);
-        if (!reader.read(res_block, *data_list_read, force_decode))
-            return false;
-        GET_METRIC(tiflash_raft_write_data_to_storage_duration_seconds, type_decode).Observe(watch.elapsedSeconds());
-        return true;
-    };
-
-    /// In TiFlash, the actions between applying raft log and schema changes are not strictly synchronized.
-    /// There could be a chance that some raft logs come after a table gets tombstoned. Take care of it when
-    /// decoding data. Check the test case for more details.
-    FAIL_POINT_PAUSE(FailPoints::pause_before_apply_raft_snapshot);
-
-    if (!atomic_decode(false))
-    {
-        GET_METRIC(tiflash_schema_trigger_count, type_raft_decode).Increment();
-        tmt.getSchemaSyncer()->syncSchemas(context);
-
-        if (!atomic_decode(true))
-            throw Exception("Pre-decode " + region->toString() + " cache to table " + std::to_string(table_id) + " block failed",
-                            ErrorCodes::LOGICAL_ERROR);
-    }
-
-    RemoveRegionCommitCache(region, *data_list_read);
-
-    return std::make_unique<RegionPreDecodeBlockData>(std::move(res_block), schema_version, std::move(*data_list_read));
 }
 
 std::tuple<TableLockHolder, std::shared_ptr<StorageDeltaMerge>, DecodingStorageSchemaSnapshotConstPtr> //
